net: ipv6: avoid nonce length overflow
Use size_t for DAD nonce option length math so len*8 does not wrap
for large option lengths while parsing and skipping data.

+++ subsys/net/ip/ipv6_nbr.c
@@ -1215,7 +1215,7 @@ static bool read_dad_nonce_opt(struct net_pkt *pkt,
 			       uint8_t len,
 			       uint8_t nonce[NET_IPV6_DAD_NONCE_LEN])
 {
-	uint8_t opt_len = len * 8U;
+	size_t opt_len = (size_t)len * 8U;
 
 	if (opt_len < NET_IPV6_DAD_NONCE_OPT_LEN) {
 		return false;
